sats kan være desimaltall
vi får simuleringer fra Simuleringstjensten hvor sats av og til er et desimaltall.

Det begynte å skje etter at vi skrudde på brukerutbetaling og vi mistenker det er relevant.

+++ sykepenger-mediators/src/main/kotlin/no/nav/helse/spleis/meldinger/model/SimuleringMessage.kt
@@ -55,7 +55,7 @@ internal class SimuleringMessage(packet: JsonMessage) : BehovMessage(packet) {
                                             refunderesOrgnummer = detalj.path("refunderesOrgNr").asText(),
                                             tilbakeføring = detalj.path("tilbakeforing").asBoolean(),
                                             sats = Simulering.Sats(
-                                                sats = detalj.path("sats").asInt(),
+                                                sats = detalj.path("sats").asDouble(),
                                                 antall = detalj.path("antallSats").asInt(),
                                                 type = detalj.path("typeSats").asText()
                                             )
