+++ sykepenger-model/src/main/kotlin/no/nav/helse/hendelser/Simulering.kt
@@ -120,7 +120,7 @@ class Simulering(
     )
 
     class Sats(
-        internal val sats: Int,
+        internal val sats: Double,
         internal val antall: Int,
         internal val type: String
     )
